@@ -1,21 +1,18 @@
 name: pop_pop
 description: Flexible and extendable components to build a bubble pop game in Flutter.
-version: 1.0.0
+version: 1.0.0-alpha.1
 homepage: https://github.com/jamieastley/pop_pop
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  rxdart: 0.24.1
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.2
-  pedantic: ^1.9.2
+  flutter_lints: ^1.0.4
 
 flutter:
